git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@2695 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CD3D9ShaderMaterialRenderer.cpp
@@ -177,7 +177,7 @@ bool CD3D9ShaderMaterialRenderer::createPixelShader(const c8* pxsh)
 	#ifdef _IRR_D3D_NO_SHADER_DEBUGGING
 
 		// compile shader without debug info
-		stubD3DXAssembleShader(pxsh, strlen(pxsh), 0, 0, 0, &code, &errors);
+		stubD3DXAssembleShader(pxsh, (UINT)strlen(pxsh), 0, 0, 0, &code, &errors);
 	#else
 
 		// compile shader and emitt some debug informations to
@@ -236,7 +236,7 @@ bool CD3D9ShaderMaterialRenderer::createVertexShader(const char* vtxsh)
 	#ifdef _IRR_D3D_NO_SHADER_DEBUGGING
 
 		// compile shader without debug info
-		stubD3DXAssembleShader(vtxsh, strlen(vtxsh), 0, 0, 0, &code, &errors);
+		stubD3DXAssembleShader(vtxsh, (UINT)strlen(vtxsh), 0, 0, 0, &code, &errors);
 
 	#else
